requirements.txt
# flask
# gunicorn
sanic
python-binance
sanic-jinja2
gunicorn
# Bybit

# To install python-binance:
# pip install pipwin
# pipwin install Twisted

#Heroku
#Git
#Heroku CLI


#webhook: https://jp-webhook.herokuapp.com/webhook



#create 15m
#create 3m
#test

# Flask Test:
# http://localhost:5000/webhook

# Flask with Redis for async coroutines?

#APScheduler

# Gunicorn - Check for memory leaks
# https://sanic.readthedocs.io/en/latest/sanic/deploying.html

# Sanic Test: http://127.0.0.1:8000/